Move JujuStepHelper to core
JujuStepHelper's main goal is to fill in where pylibjuju is lacking by
using the juju cli.
