+++ sunbeam-python/sunbeam/core/juju.py
@@ -19,6 +19,8 @@ import base64
 import json
 import logging
 import os
+import subprocess
+import tempfile
 import typing
 from datetime import datetime
 from pathlib import Path
@@ -47,6 +49,8 @@ from juju.errors import (
 from juju.machine import Machine
 from juju.model import Model
 from juju.unit import Unit
+from packaging import version
+from snaphelpers import Snap
 
 from sunbeam.clusterd.client import Client
 from sunbeam.core.common import SunbeamException
@@ -1268,3 +1272,342 @@ class JujuHelper:
             await facade.MergeBindings(request)
         except JujuError as e:
             raise JujuException(f"Failed to merge bindings: {str(e)}") from e
+
+
+class JujuStepHelper:
+    jhelper: JujuHelper
+
+    def _get_juju_binary(self) -> str:
+        """Get juju binary path."""
+        snap = Snap()
+        juju_binary = snap.paths.snap / "juju" / "bin" / "juju"
+        return str(juju_binary)
+
+    def _juju_cmd(self, *args):
+        """Runs the specified juju command line command.
+
+        The command will be run using the json formatter. Invoking functions
+        do not need to worry about the format or the juju command that should
+        be used.
+
+        For example, to run the juju bootstrap k8s, this method should
+        be invoked as:
+
+          self._juju_cmd('bootstrap', 'k8s')
+
+        Any results from running with json are returned after being parsed.
+        Subprocess execution errors are raised to the calling code.
+
+        :param args: command to run
+        :return:
+        """
+        cmd = [self._get_juju_binary()]
+        cmd.extend(args)
+        cmd.extend(["--format", "json"])
+
+        LOG.debug(f'Running command {" ".join(cmd)}')
+        process = subprocess.run(cmd, capture_output=True, text=True, check=True)
+        LOG.debug(f"Command finished. stdout={process.stdout}, stderr={process.stderr}")
+
+        return json.loads(process.stdout.strip())
+
+    def check_model_present(self, model_name) -> bool:
+        """Determines if the step should be skipped or not.
+
+        :return: True if the Step should be skipped, False otherwise
+        """
+        try:
+            run_sync(self.jhelper.get_model(model_name))
+            return True
+        except ModelNotFoundException:
+            LOG.debug(f"Model {model_name} not found")
+            return False
+
+    def get_clouds(
+        self, cloud_type: str, local: bool = False, controller: str | None = None
+    ) -> list:
+        """Get clouds based on cloud type.
+
+        If local is True, return clouds registered in client.
+        If local is False, return clouds registered in client and controller.
+        If local is False and controller specified, return clouds registered
+        in controller.
+        """
+        clouds = []
+        cmd = ["clouds"]
+        if local:
+            cmd.append("--client")
+        else:
+            if controller:
+                cmd.extend(["--controller", controller])
+        clouds_from_juju_cmd = self._juju_cmd(*cmd)
+        LOG.debug(f"Available clouds in juju are {clouds_from_juju_cmd.keys()}")
+
+        for name, details in clouds_from_juju_cmd.items():
+            if details["type"] == cloud_type:
+                clouds.append(name)
+
+        LOG.debug(f"There are {len(clouds)} {cloud_type} clouds available: {clouds}")
+
+        return clouds
+
+    def get_credentials(
+        self, cloud: str | None = None, local: bool = False
+    ) -> dict[str, dict]:
+        """Get credentials."""
+        cmd = ["credentials"]
+        if local:
+            cmd.append("--client")
+        if cloud:
+            cmd.append(cloud)
+        return self._juju_cmd(*cmd)
+
+    def get_controllers(self, clouds: list | None = None) -> list:
+        """Get controllers hosted on given clouds.
+
+        if clouds is None, return all the controllers.
+        """
+        controllers = self._juju_cmd("controllers")
+        controllers = controllers.get("controllers", {}) or {}
+        if clouds is None:
+            return list(controllers.keys())
+
+        existing_controllers = [
+            name for name, details in controllers.items() if details["cloud"] in clouds
+        ]
+        LOG.debug(
+            f"There are {len(existing_controllers)} existing {clouds} "
+            f"controllers running: {existing_controllers}"
+        )
+        return existing_controllers
+
+    def get_external_controllers(self) -> list:
+        """Get all external controllers registered."""
+        snap = Snap()
+        data_location = snap.paths.user_data
+        external_controllers = []
+
+        controllers = self.get_controllers()
+        for controller in controllers:
+            account_file = data_location / f"{controller}.yaml"
+            if account_file.exists():
+                external_controllers.append(controller)
+
+        return external_controllers
+
+    def get_controller(self, controller: str) -> dict:
+        """Get controller definition."""
+        try:
+            return self._juju_cmd("show-controller", controller)[controller]
+        except subprocess.CalledProcessError as e:
+            LOG.debug(e)
+            raise ControllerNotFoundException() from e
+
+    def add_cloud(self, name: str, cloud: dict, controller: str | None) -> bool:
+        """Add cloud to client clouds.
+
+        If controller is specified, add cloud to both client
+        and given controller.
+        """
+        if cloud["clouds"][name]["type"] not in ("manual", "maas"):
+            return False
+
+        with tempfile.NamedTemporaryFile() as temp:
+            temp.write(yaml.dump(cloud).encode("utf-8"))
+            temp.flush()
+            cmd = [
+                self._get_juju_binary(),
+                "add-cloud",
+                name,
+                "--file",
+                temp.name,
+                "--client",
+            ]
+            if controller:
+                cmd.extend(["--controller", controller, "--force"])
+            LOG.debug(f'Running command {" ".join(cmd)}')
+            process = subprocess.run(cmd, capture_output=True, text=True, check=True)
+            LOG.debug(
+                f"Command finished. stdout={process.stdout}, stderr={process.stderr}"
+            )
+
+        return True
+
+    def add_credential(self, cloud: str, credential: dict, controller: str | None):
+        """Add credential to client credentials."""
+        with tempfile.NamedTemporaryFile() as temp:
+            temp.write(yaml.dump(credential).encode("utf-8"))
+            temp.flush()
+            cmd = [
+                self._get_juju_binary(),
+                "add-credential",
+                cloud,
+                "--file",
+                temp.name,
+                "--client",
+            ]
+            if controller:
+                cmd.extend(["--controller", controller])
+            LOG.debug(f'Running command {" ".join(cmd)}')
+            process = subprocess.run(cmd, capture_output=True, text=True, check=True)
+            LOG.debug(
+                f"Command finished. stdout={process.stdout}, stderr={process.stderr}"
+            )
+
+    def integrate(
+        self,
+        model: str,
+        provider: str,
+        requirer: str,
+        ignore_error_if_exists: bool = True,
+    ):
+        """Juju integrate applications."""
+        cmd = [
+            self._get_juju_binary(),
+            "integrate",
+            "-m",
+            model,
+            provider,
+            requirer,
+        ]
+        try:
+            LOG.debug(f'Running command {" ".join(cmd)}')
+            process = subprocess.run(cmd, capture_output=True, text=True, check=True)
+            LOG.debug(
+                f"Command finished. stdout={process.stdout}, stderr={process.stderr}"
+            )
+        except subprocess.CalledProcessError as e:
+            LOG.debug(e.stderr)
+            if ignore_error_if_exists and "already exists" not in e.stderr:
+                raise e
+
+    def remove_relation(self, model: str, provider: str, requirer: str):
+        """Juju remove relation."""
+        cmd = [
+            self._get_juju_binary(),
+            "remove-relation",
+            "-m",
+            model,
+            provider,
+            requirer,
+        ]
+        LOG.debug(f'Running command {" ".join(cmd)}')
+        process = subprocess.run(cmd, capture_output=True, text=True, check=True)
+        LOG.debug(f"Command finished. stdout={process.stdout}, stderr={process.stderr}")
+
+    def revision_update_needed(
+        self, application_name: str, model: str, status: dict | None = None
+    ) -> bool:
+        """Check if a revision update is available for an applicaton.
+
+        :param application_name: Name of application to check for updates for
+        :param model: Model application is in
+        :param status: Dictionay of model status
+        """
+        if not status:
+            status = run_sync(self.jhelper.get_model_status_full(model))
+        app_status = status["applications"].get(application_name, {})
+        if not app_status:
+            LOG.debug(f"{application_name} not present in model")
+            return False
+        deployed_revision = int(self._extract_charm_revision(app_status["charm"]))
+        charm_name = self._extract_charm_name(app_status["charm"])
+        deployed_channel = self.normalise_channel(app_status["charm-channel"])
+        if len(deployed_channel.split("/")) > 2:
+            LOG.debug(f"Cannot calculate upgrade for {application_name}, branch in use")
+            return False
+        available_revision = run_sync(
+            self.jhelper.get_available_charm_revision(
+                model, charm_name, deployed_channel
+            )
+        )
+        return bool(available_revision > deployed_revision)
+
+    def get_charm_deployed_versions(self, model: str) -> dict:
+        """Return charm deployed info for all the applications in model.
+
+        For each application, return a tuple of charm name, channel and revision.
+        Example output:
+        {"keystone": ("keystone-k8s", "2023.2/stable", 234)}
+        """
+        status = run_sync(self.jhelper.get_model_status_full(model))
+
+        apps = {}
+        for app_name, app_status in status.get("applications", {}).items():
+            charm_name = self._extract_charm_name(app_status["charm"])
+            deployed_channel = self.normalise_channel(app_status["charm-channel"])
+            deployed_revision = int(self._extract_charm_revision(app_status["charm"]))
+            apps[app_name] = (charm_name, deployed_channel, deployed_revision)
+
+        return apps
+
+    def get_apps_filter_by_charms(self, model: str, charms: list) -> list:
+        """Return apps filtered by given charms.
+
+        Get all apps from the model and return only the apps deployed with
+        charms in the provided list.
+        """
+        deployed_all_apps = self.get_charm_deployed_versions(model)
+        return [
+            app_name
+            for app_name, (charm, channel, revision) in deployed_all_apps.items()
+            if charm in charms
+        ]
+
+    def normalise_channel(self, channel: str) -> str:
+        """Expand channel if it is using abbreviation.
+
+        Juju supports abbreviating latest/{risk} to {risk}. This expands it.
+
+        :param channel: Channel string to normalise
+        """
+        if channel in ["stable", "candidate", "beta", "edge"]:
+            channel = f"latest/{channel}"
+        return channel
+
+    def _extract_charm_name(self, charm_url: str) -> str:
+        """Extract charm name from charm url.
+
+        :param charm_url: Url to examine
+        """
+        # XXX There must be a better way. ch:amd64/jammy/cinder-k8s-50 -> cinder-k8s
+        return charm_url.split("/")[-1].rsplit("-", maxsplit=1)[0]
+
+    def _extract_charm_revision(self, charm_url: str) -> str:
+        """Extract charm revision from charm url.
+
+        :param charm_url: Url to examine
+        """
+        return charm_url.split("-")[-1]
+
+    def channel_update_needed(self, channel: str, new_channel: str) -> bool:
+        """Compare two channels and see if the second is 'newer'.
+
+        :param current_channel: Current channel
+        :param new_channel: Proposed new channel
+        """
+        risks = ["stable", "candidate", "beta", "edge"]
+        current_channel = self.normalise_channel(channel)
+        current_track, current_risk = current_channel.split("/")
+        new_track, new_risk = new_channel.split("/")
+        if current_track != new_track:
+            try:
+                return version.parse(current_track) < version.parse(new_track)
+            except version.InvalidVersion:
+                LOG.error("Error: Could not compare tracks")
+                return False
+        if risks.index(current_risk) < risks.index(new_risk):
+            return True
+        else:
+            return False
+
+    def get_model_name_with_owner(self, model: str) -> str:
+        """Return model name with owner name.
+
+        :param model: Model name
+
+        Raises ModelNotFoundException if model does not exist.
+        """
+        model_with_owner = run_sync(self.jhelper.get_model_name_with_owner(model))
+
+        return model_with_owner
