+++ sunbeam-python/sunbeam/features/pro/feature.py
@@ -27,7 +27,7 @@ from snaphelpers import Snap
 from sunbeam.clusterd.client import Client
 from sunbeam.core.common import BaseStep, Result, ResultType, run_plan
 from sunbeam.core.deployment import Deployment
-from sunbeam.core.juju import JujuHelper, TimeoutException, run_sync
+from sunbeam.core.juju import JujuHelper, JujuStepHelper, TimeoutException, run_sync
 from sunbeam.core.manifest import Manifest, SoftwareConfig, TerraformManifest
 from sunbeam.core.terraform import (
     TerraformException,
@@ -35,7 +35,6 @@ from sunbeam.core.terraform import (
     TerraformInitStep,
 )
 from sunbeam.features.interface.v1.base import EnableDisableFeature
-from sunbeam.steps.juju import JujuStepHelper
 from sunbeam.utils import argument_with_deprecated_option
 
 LOG = logging.getLogger(__name__)
